+++ .gitignore
@@ -14,7 +14,6 @@ dist/
 downloads/
 eggs/
 .eggs/
-lib/
 lib64/
 parts/
 sdist/
@@ -108,3 +107,7 @@ venv.bak/
 
 # PyCharm
 .idea/
+
+# Serverless
+.serverless/
+_build/
