+++ pierre/requirements.txt
@@ -1,0 +1,1 @@
+requests==2.13.0
