+++ requirements-dev.txt
@@ -1,0 +1,3 @@
+pytest==3.0.7
+pytest-cov==2.4.0
+mock==2.0.0
